Make configuration consistent
***BREAKING CHANGE***
The couch database client wrapper no longer expects the server
configuration to be wrapped in an object with the pertinent
configuration to be found at the `db_config` namespace.  You are now
expected to submit an object with all of the configuration parameters
(`url`, `port`, `username`, `password`) at the top level of the object.